--- Assets/Audio/FireShoot2.ogg.meta
@@ -1,21 +1,0 @@
-fileFormatVersion: 2
-guid: ca679721d8a1b6243acc9a7ce566789b
-timeCreated: 1430492558
-licenseType: Store
-AudioImporter:
-  serializedVersion: 6
-  defaultSettings:
-    loadType: 0
-    sampleRateSetting: 0
-    sampleRateOverride: 44100
-    compressionFormat: 1
-    quality: 1
-    conversionMode: 0
-  platformSettingOverrides: {}
-  forceToMono: 1
-  preloadAudioData: 1
-  loadInBackground: 0
-  3D: 1
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
