+++ Assets/Audio/PlayerHabilities.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a3cc8026ecb673d43a01d32c0e11d506
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
