--- Assets/Particles/UsedParticles/Healing circle.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 12a56ee4630040c4eb17ae9707ed4cef
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
