--- Assets/Prefabs/FireballOne.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 9d4c32ae21e86754092ba2e90b16bb41
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
