--- Assets/Prefabs/Meteors AOE.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 4a5819bd580894245beba72b9ee9aca1
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
