+++ Assets/Prefabs/PlayerHabilities.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 024224f7f3e078e4db0cb08e8866f2c6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
