+++ Assets/Prefabs/UIPrefabs/Hud.prefab
@@ -1798,16 +1798,16 @@ MonoBehaviour:
           m_BoolArgument: 1
         m_CallState: 2
       - m_Target: {fileID: 0}
-        m_TargetAssemblyTypeName: InGameOptionsController, Assembly-CSharp
-        m_MethodName: ShowOptionsInGame
-        m_Mode: 1
+        m_TargetAssemblyTypeName: UnityEngine.GameObject, UnityEngine
+        m_MethodName: SetActive
+        m_Mode: 6
         m_Arguments:
           m_ObjectArgument: {fileID: 0}
           m_ObjectArgumentAssemblyTypeName: UnityEngine.Object, UnityEngine
           m_IntArgument: 0
           m_FloatArgument: 0
           m_StringArgument: 
-          m_BoolArgument: 0
+          m_BoolArgument: 1
         m_CallState: 2
 --- !u!1 &7659870808980052729
 GameObject:
@@ -4070,4 +4070,3 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 2789280d48aac654287fec0e1742a2ae, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  m_defeatedEnemies: 0
