--- Assets/Scripts/Portal.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3dd960247d25cd74991513765d23ad46
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
