+++ Assets/Scripts/Portal.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 82cb8d5b6121240459075c30c059548d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
