--- Assets/Scripts/PortalButton.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3706c25b588ddac4a9831e8c6778da5e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
